Resolved CID 351586 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument sb.st_size to format specifier %u was expected to have type unsigned int but has type long

+++ src/logonserver/Auth/AutoPatcher.cpp
@@ -157,7 +157,7 @@ void PatchMgr::initialize()
             continue;
         }
 
-        LogNotice("PatchMgr : Found patch for b%u locale `%s` (%u bytes).", srcversion, locality, sb.st_size);
+        LogNotice("PatchMgr : Found patch for b%u locale `%s` (%u bytes).", srcversion, locality, static_cast<uint32_t>(sb.st_size));
         pPatch = new Patch;
         size = sb.st_size;
         pPatch->FileSize = size;
